+++ src/main/java/com/fvukic/webshop/service/implementation/OrderServiceImplementation.java
@@ -48,6 +48,9 @@ public class OrderServiceImplementation implements OrderService {
 
     @Override
     public void deleteOrderRequest(Integer orderId) {
+        if (orderRepository.existsById(orderId)){
+            throw new EntityWithIdNotFoundException(ErrorResponse.ERROR_WRONG_ID,orderId);
+        }
         orderRepository.deleteById(orderId);
     }
